requirements.txt
# LangGraph y LangChain
langgraph==0.2.34
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16

# Base de datos
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Análisis de datos
pandas==2.1.4
numpy==1.24.4

# Utilidades
python-dateutil==2.8.2
python-dotenv==1.0.0
pydantic==2.5.2

# Desarrollo y testing
pytest==7.4.3
black==23.12.1
flake8==6.1.0